--- src/main/java/alpaca/Error.java
@@ -1,3 +1,0 @@
-public class Error {
-
-}
