+++ src/main/java/alpaca/Parser.java
@@ -1,5 +1,7 @@
 package alpaca;
 
+import alpaca.command.*;
+
 public class Parser {
     public static Command parse(String input) {
         String[] words = input.split(" ", 2);
