+++ src/main/java/alpaca/command/ListCommand.java
@@ -1,4 +1,8 @@
-package alpaca;
+package alpaca.command;
+
+import alpaca.Storage;
+import alpaca.TaskList;
+import alpaca.Ui;
 
 public class ListCommand extends Command {
     @Override
