+++ src/main/java/alpaca/task/Deadline.java
@@ -1,4 +1,5 @@
-package alpaca;
+package alpaca.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
