+++ src/main/java/alpaca/task/Event.java
@@ -1,4 +1,5 @@
-package alpaca;
+package alpaca.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
