+++ src/main/java/alpaca/task/Todo.java
@@ -1,4 +1,4 @@
-package alpaca;
+package alpaca.task;
 
 public class Todo extends Task {
     public Todo(String description) {
